refactor: rename application package
