+++ src/main/java/com/vegan/mealplanner/VeganMealPlannerApiApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.vegan.mealplanner;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
