+++ src/test/java/com/vegan/mealplanner/VeganMealPlannerApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.vegan.mealplanner;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
